+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `:parallel` option to `start` and `stop` options. Value is number of threads to use, to start/stop independent states in parallel.
 - Internally mount-lite now builds a proper graph of dependencies, instead of a sequence.
 - Add processing of `:dependencies` meta data in a state var, in order to influence the deduced state dependency graph.
+- Add `status` function to main API.
 
 ### Changed
 - Start/stop option `:up-to` now uses the internal dependency graph to only start (or stop) the dependencies (or dependents) of the given var.